get evaluator out of parser.rs
