+++ src/interpreter/mod.rs
@@ -3,5 +3,7 @@ pub mod basic;
 pub mod lexer;
 pub mod parser;
 pub mod evaluator;
+pub mod symbol;
+pub mod node;
 
-pub use self::{lexer::*, parser::*, basic::*, evaluator::*};
+pub use self::{lexer::*, parser::*, basic::*, evaluator::*, node::*};
